add query18
this includes adding customer/orders table into exp scripts

+++ tpch-2.14.0/experiments/final-schema/orders-enc.sql
@@ -4,7 +4,7 @@ CREATE TABLE orders_enc  (
     o_orderstatus_DET    tinyint unsigned not null,
     o_totalprice_DET     bigint(20) unsigned not null,
     o_totalprice_OPE     binary(16) not null,
-    o_orderdate_DET      integer unsigned not null,
+    o_orderdate_DET      mediumint unsigned not null,
     o_orderdate_OPE      bigint(20) unsigned not null,
     o_orderpriority_DET  binary(15) not null,  
     o_clerk_DET          binary(15) not null,
